+++ app/src/main/java/org/simple/clinic/home/overdue/search/OverdueSearchScreen.kt
@@ -11,7 +11,6 @@ import android.view.ViewGroup
 import android.view.inputmethod.EditorInfo
 import android.widget.ArrayAdapter
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.lifecycleScope
 import androidx.paging.LoadState
 import androidx.paging.PagingData
 import androidx.recyclerview.widget.RecyclerView
@@ -192,8 +191,7 @@ class OverdueSearchScreen : BaseScreen<
 
   override fun createEffectHandler(viewEffectsConsumer: Consumer<OverdueSearchViewEffect>) = effectHandlerFactory
       .create(
-          viewEffectsConsumer = viewEffectsConsumer,
-          pagingCacheScope = viewLifecycleOwner.lifecycleScope
+          viewEffectsConsumer = viewEffectsConsumer
       )
       .build()
 
@@ -300,7 +298,6 @@ class OverdueSearchScreen : BaseScreen<
   }
 
   override fun hideSearchResults() {
-    overdueSearchListAdapter.submitData(lifecycle, PagingData.empty())
     overdueSearchRecyclerView.visibility = GONE
   }