+++ app/src/main/java/org/simple/clinic/home/overdue/search/OverdueSearchUiRenderer.kt
@@ -48,8 +48,6 @@ class OverdueSearchUiRenderer(
         // No-op
       }
     }
-
-    ui.setOverdueSearchResultsPagingData(model.overdueSearchResults, model.selectedOverdueAppointments, model.searchQuery.orEmpty())
   }
 
   private fun renderOverdueListSelectedCount(selectedOverdueAppointments: Set<UUID>) {