+++ app/src/main/java/org/simple/clinic/home/overdue/search/OverdueSearchViewEffectHandler.kt
@@ -16,6 +16,11 @@ class OverdueSearchViewEffectHandler(
       OpenSelectShareFormatDialog -> uiActions.openSelectShareFormatDialog()
       OpenShareInProgressDialog -> uiActions.openShareInProgressDialog()
       ShowNoInternetConnectionDialog -> uiActions.showNoInternetConnectionDialog()
+      is SetOverdueSearchPagingData -> uiActions.setOverdueSearchResultsPagingData(
+          overdueSearchResults = viewEffect.overdueSearchResults,
+          selectedOverdueAppointments = viewEffect.selectedOverdueAppointments,
+          searchQuery = viewEffect.searchQuery
+      )
     }.exhaustive()
   }
 }