+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Don't show search history when overdue search v2 is enabled
 - Bump room to 2.4.3
 - Remove overdue search v2 feature flag key
+- Fix overdue search not searching on query changes
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-07-25-8342